% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans_calc.R
\name{calculate_kmeans}
\alias{calculate_kmeans}
\title{Calculate K-Means Clusters}
\usage{
calculate_kmeans(data, centers)
}
\arguments{
\item{data}{A numeric matrix or data frame.}

\item{centers}{The number of clusters.}
}
\value{
A list containing the K-means result.
}
\description{
This function calculates K-means clusters for a given dataset.
}
