% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans_plot.R
\name{plot_kmeans}
\alias{plot_kmeans}
\title{Plot K-Means Clusters}
\usage{
plot_kmeans(kmeans_result, data)
}
\arguments{
\item{kmeans_result}{The result object from \code{calculate_kmeans}.}

\item{data}{The original dataset used for clustering.}
}
\value{
A ggplot object.
}
\description{
This function plots the results of a K-means clustering.
}
